+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-Nothing yet
+### Added
+- [#66](https://github.com/Kashoo/synctos/issues/66): Modular document definition files
 
 ## [1.5.0] - 2016-12-14
 ### Added